Patch for minor issue #768
This line of code doens't work on Windows and is not needed. Plus, SOPC is not available since Quartus II. Probably this line could be completely removed.

+++ hdl/quartus/build_bladerf.sh
@@ -470,5 +470,7 @@ echo ""
 printf "%s %02d:%02d:%02d\n" "Total Build Time:" "$(($SECONDS / 3600))" "$((($SECONDS / 60) % 60))" "$(($SECONDS % 60))"
 echo ""
 
-# Delete empty SOPC directories in the user's home directory
-find ~ -maxdepth 1 -type d -empty -iname "sopc_altera_pll*" -delete
+if [ "x$(uname)" != "xLinux" ]; then
+    # Delete empty SOPC directories in the user's home directory
+    find ~ -maxdepth 1 -type d -empty -iname "sopc_altera_pll*" -delete
+fi
